struct Product added

+++ NFCReadDemo-iOS-SwiftUI/model_classes/Product.swift
@@ -12,9 +12,9 @@ struct Product: Codable {
   var id                 : Int?      = nil
   var title              : String?   = nil
   var description        : String?   = nil
-  var price              : Int?      = nil
+  var price              : Double?      = nil
   var discountPercentage : Double?   = nil
-  var rating             : Int?      = nil
+  var rating             : Double?      = nil
   var stock              : Int?      = nil
   var brand              : String?   = nil
   var category           : String?   = nil
@@ -34,25 +34,22 @@ struct Product: Codable {
     case category           = "category"
     case thumbnail          = "thumbnail"
     case images             = "images"
-  
   }
 
-  init(from decoder: Decoder) throws {
-    let values = try decoder.container(keyedBy: CodingKeys.self)
-
-    id                 = try values.decodeIfPresent(Int.self      , forKey: .id                 )
-    title              = try values.decodeIfPresent(String.self   , forKey: .title              )
-    description        = try values.decodeIfPresent(String.self   , forKey: .description        )
-    price              = try values.decodeIfPresent(Int.self      , forKey: .price              )
-    discountPercentage = try values.decodeIfPresent(Double.self   , forKey: .discountPercentage )
-    rating             = try values.decodeIfPresent(Int.self      , forKey: .rating             )
-    stock              = try values.decodeIfPresent(Int.self      , forKey: .stock              )
-    brand              = try values.decodeIfPresent(String.self   , forKey: .brand              )
-    category           = try values.decodeIfPresent(String.self   , forKey: .category           )
-    thumbnail          = try values.decodeIfPresent(String.self   , forKey: .thumbnail          )
-    images             = try values.decodeIfPresent([String].self , forKey: .images             )
- 
-  }
+    init(from decoder: Decoder) throws {
+        let values = try decoder.container(keyedBy: CodingKeys.self)
+        id = try values.decodeIfPresent(Int.self, forKey: .id)
+        title = try values.decodeIfPresent(String.self, forKey: .title)
+        description = try values.decodeIfPresent(String.self, forKey: .description)
+        price = try values.decodeIfPresent(Double.self, forKey: .price)
+        discountPercentage = try values.decodeIfPresent(Double.self, forKey: .discountPercentage)
+        rating = try values.decodeIfPresent(Double.self, forKey: .rating)
+        stock = try values.decodeIfPresent(Int.self, forKey: .stock)
+        brand = try values.decodeIfPresent(String.self, forKey: .brand)
+        category = try values.decodeIfPresent(String.self, forKey: .category)
+        thumbnail = try values.decodeIfPresent(String.self, forKey: .thumbnail)
+        images = try values.decodeIfPresent([String].self, forKey: .images)
+    }
 
   init() {
